Added a ItemViewModel class, this class holds a list of modles as properties.
I then use this modle in the home controler async task to create database lists one at a time via an await request.

Then the lists are used to populate an instance of the ItemViewModel. This ItemViewModel then gets returned to the about page.

I belive that this will allow us to use more than one model in a view. Further testing will be needed.

for more info see this site http://www.c-sharpcorner.com/UploadFile/ff2f08/multiple-models-in-single-view-in-mvc/